Check requirements before state machine
