+++ docs/index.rst
@@ -96,6 +96,7 @@ This Github Repo_ contains logo images. If you are using TRD and want to let eve
    run
    plugins
    linuxservice
+   state_machine
    contributors
    testing
    statistics
