+++ docs/state_machine.rst
@@ -1,0 +1,20 @@
+TRD State Machines
+====================
+
+TRD is designed as a state machine to make sure it's reliable.
+
+In the future TRD will be `formally verified`_ to avoid undetected bugs.
+
+Launch state machine diagram
+--------------------------------
+
+.. image:: fsm/graphviz/launch_state_diagram.png
+  :alt: Launch State Diagram
+
+Configuration state machine diagram
+-------------------------------------
+
+.. image:: fsm/graphviz/launch_state_diagram.png
+  :alt: Configuration State Diagram
+
+.. _formally verified: https://en.wikipedia.org/wiki/Formal_verification
